Complete Phase 2: Module Refactoring
Split three oversized modules into focused submodules, eliminating all
module size violations while maintaining 100% test compatibility.

Changes:

1. **Interpreter Module** (848 lines → 3 submodules)
   - src/interpreter/mod.rs (496 lines) - Core interpreter and public API
   - src/interpreter/preamble.rs (268 lines) - Preamble building for composition
   - src/interpreter/shell.rs (199 lines) - Shell command execution

2. **Parser Module** (694 lines → 4 submodules)
   - src/parser/mod.rs (501 lines) - Main parsing logic and tests
   - src/parser/attributes.rs (162 lines) - Attribute parsing
   - src/parser/preprocessing.rs (29 lines) - Input preprocessing
   - src/parser/shebang.rs (23 lines) - Shebang detection

3. **MCP Module** (669 lines → 4 submodules)
   - src/mcp/mod.rs (258 lines) - Protocol handling
   - src/mcp/tools.rs (151 lines) - Tool schema and inspection
   - src/mcp/handlers.rs (149 lines) - JSON-RPC request handlers
   - src/mcp/mapping.rs (131 lines) - Argument mapping

Benefits:
- ✅ All modules now under or near 500-line guideline
- ✅ Improved code organization with single responsibility per file
- ✅ Better maintainability and easier navigation
- ✅ Clear API boundaries with pub(super) visibility
- ✅ No breaking changes to public API

Results:
- All 156 tests passing
- Zero module size violations (was 3)
- Successful compilation with minimal warnings
- Maintained backward compatibility

Documentation:
- Added docs/PHASE2_COMPLETE.md with detailed breakdown

Next: Phase 3 - Function Extraction (5 functions >50 lines)
